framework/requirements.txt
pycryptodome
cryptography
pyyaml
tabulate
pyopenssl